StateMachineActions, Repeat: rewrite to use Generator
One pro of this method is the ability to use filters to weed out
actions that cannot be executed *before* running them. This should
reduce the number of actions executed and lower the useless noise
in the logs - instead of executing r.Skip() in every action, a filter
gets rid of it entirely.